packages/http/httpx/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.poetry]
name = "microsoft-kiota-http"
version = "0.0.4"
description = "Core abstractions for kiota generated libraries in Python"
authors = ["Microsoft <graphtooling+python@microsoft.com>"]
license = "MIT"
readme = "README.md"
keywords = ["kiota", "openAPI", "Microsoft", "Graph"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
homepage = "https://github.com/microsoft/kiota#readme"
repository = "https://github.com/microsoft/kiota-python"
documentation = "https://microsoft.github.io/kiota/"
packages = [{include = "kiota_http"}]	

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
microsoft-kiota-abstractions = ">=1.0.0,<2.0.0"
opentelemetry-api = ">=1.20.0"
opentelemetry-sdk = ">=1.20.0"
httpx = {extras = ["http2"], version = "^0.23.0"}

[tool.poetry.group.dev.dependencies]
yapf = "^0.40.2"
isort = "^5.13.2"
pylint = "^3.2.7"
mypy = "^1.11.2"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"

[tool.mypy]
warn_unused_configs = true
files = "kiota_http"
ignore_missing_imports = true

[tool.yapf]
based_on_style = "pep8"
dedent_closing_brackets = true
each_dict_entry_on_separate_line = true
column_limit = 100

[tool.isort]
profile = "hug"